requirements.txt
streamlit==0.89.0
pandas==1.3.3
streamlit_echarts==0.1.0
typing-extensions==3.10.0.2
sqlalchemy==1.4.25
dataclasses==0.6
python-dotenv==0.18.0
SpeechRecognition==3.8.1
pyttsx3==2.90
